+++ eRestaurant/Controllers/Master/VarientController.cs
@@ -14,21 +14,24 @@ namespace RocketPOS.Controllers.Master
 {
     public class VarientController : Controller
     {
+        private readonly ICommonService _iCommonService;
         private readonly IVarientService _iVarientService;
         private readonly IDropDownService _iDropDownService;
         private IStringLocalizer<RocketPOSResources> _sharedLocalizer;
         private LocService _locService;
 
-        public VarientController(IVarientService varientService, IDropDownService idropDownService, IStringLocalizer<RocketPOSResources> sharedLocalizer, LocService locService)
+        public VarientController(IVarientService varientService, ICommonService iCommonService, IDropDownService idropDownService, IStringLocalizer<RocketPOSResources> sharedLocalizer, LocService locService)
         {
             _iVarientService = varientService;
             _iDropDownService = idropDownService;
+            _iCommonService = iCommonService;
             _sharedLocalizer = sharedLocalizer;
             _locService = locService;
         }
 
         public ActionResult Index()
         {
+            _iCommonService.GetPageWiseRoleRigths("Varient");
             List<VarientModel> varientModel = new List<VarientModel>();
             varientModel = _iVarientService.GetVarientList().ToList();
             return View(varientModel);
@@ -37,13 +40,20 @@ namespace RocketPOS.Controllers.Master
         public ActionResult Varient(int? id)
         {
             VarientModel varientModel = new VarientModel();
-            if (id > 0)
+            if (UserRolePermissionForPage.Add == true || UserRolePermissionForPage.Edit == true)
             {
-                int varientId = Convert.ToInt32(id);
-                varientModel = _iVarientService.GetAddonesById(varientId);
+                if (id > 0)
+                {
+                    int varientId = Convert.ToInt32(id);
+                    varientModel = _iVarientService.GetAddonesById(varientId);
+                }
+                varientModel.FoodMenuList = _iDropDownService.GetFoodMenuList();
+                return View(varientModel);
             }
-            varientModel.FoodMenuList = _iDropDownService.GetFoodMenuList();
-            return View(varientModel);
+            else
+            {
+                return RedirectToAction("NotFound", "Error");
+            }
         }
 
         [HttpPost]
@@ -72,15 +82,22 @@ namespace RocketPOS.Controllers.Master
                 var result = _iVarientService.InsertVarient(varientModel);
                 ViewBag.Result = _locService.GetLocalizedHtmlString("SaveSuccess");
             }
- 
+
             return RedirectToAction("Index", "Varient");
         }
 
         public ActionResult Delete(int id)
         {
-            var deletedid = _iVarientService.DeleteVarient(id);
+            if (UserRolePermissionForPage.Delete == true)
+            {
+                var deletedid = _iVarientService.DeleteVarient(id);
 
-            return RedirectToAction(nameof(Index));
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                return RedirectToAction("NotFound", "Error");
+            }
         }
 
         private string ValidationVarient(VarientModel varientModel)
